+++ packages/gdu/commands/start/runSPA.ts
@@ -40,7 +40,7 @@ export const runSPA = async (guruConfig: GuruConfig, isDebug) => {
 
 	// eslint-disable-next-line unicorn/prefer-prototype-methods
 	const webpackConfig: Configuration = hooks.webpackConfig
-		.call(webpackConfigs(appEnv, isDebug, null, guruConfig?.standalone))
+		.call(webpackConfigs(appEnv, isDebug, guruConfig?.standalone))
 		.find(({ name }) => name === appEnv);
 
 	const consumerHtmlTemplate = getConsumerHtmlTemplate(guruConfig);
@@ -48,7 +48,6 @@ export const runSPA = async (guruConfig: GuruConfig, isDebug) => {
 	webpackConfig.plugins.push(
 		new HtmlWebpackPlugin({
 			template: consumerHtmlTemplate ?? 'auto',
-			// @ts-ignore
 			scriptLoading: 'module',
 		}),
 	);